+++ was/src/main/java/com/senifit/was/repository/center/CentersRepository.java
@@ -3,9 +3,10 @@ package com.senifit.was.repository.center;
 import com.senifit.was.entity.Centers;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+import java.util.Optional;
 
 @Repository
 public interface CentersRepository extends JpaRepository<Centers, Long> {
     boolean existsByid(String id); // 아이디 중복확인에 사용
-    Centers findByid(String id);
+    Optional<Centers> findByid(String id);
 }
